[ichthyop 0003706]: Advanced parameter editor get frozen
The loading of the parameter editors (advanced and user-friendly) was done in the EDT (Event Dispatch Thread, the thread dedicated to the graphical events). It has no consequence on 90% of the case, because it is displayed quickly, but the user-interface got frozen when, for example, the xml configuration file is read through the network.
To avoid the saturation of the EDT, I now load the parameter editor with a Task object that is thread safe.