[ichthyop 0003717]: Export simulation outputs into GoogleEarth format
(WMSMapper.java)
Created some methods to convert the NetCDF output files into KMZ file. I used the excellent JAK library http://labs.micromata.de/display/jak/Home from Micromata.

(SFTask.java)
SFTask for Success / Failure task. An extension of the Task object that fires change of "succeeded" and "failed" properties for making the display of the green/red flag in the status bar easier.
Overrided the setMessage(String message, boolen persistent) method to determine whether the application should delete the message after a few seconds or on the contrary let it visible until a new message is sent.

(JStatusBar.java)
Listent to the new management of the messages : sometimes it has to reset the message label, sometimes not.

(IchthyopView.java)
Added a Task and button "Export to KMZ".
Replaced all the Task by the new object SFTask.

(LoggerScrollPane.java)
It is now connected to the Task monitor to listen the messages sent by the application when executing some tasks.