[ichthyop 0003704]: Zones are not displayed in the simulation preview
(SimulationUI.java)
the CellUI.getColor was only checking at the depth for setting the color. It now detects whether the cell is within a zone or not.

(ZoneManager.java)
the getZones(typeZone) only returns the enabled zone. By default they were all enabled but it is now something to set by the user or at least by the configuration editor when choosing to enable recruitment or zone release.

Made changes in other classes to make sure that the ZoneManager.getZones(typeZone) method will not cause NullPointerException in classes that call it.