(OutputNC.java) - by David K. Small modification to OutputNC.java so that while it is writing out to a netCDF file, it uses a filename of the form FILENAME.nc.part.  When it is completely finished writing out to the file (at OutputNC.close()), the file is renamed to FILENAME.nc
This is useful for separating incomplete from completed netCDF output files.
The use case I am interested in is for running multiple instances of 
ichthyop on a cluster, each of which generates many large output files, but
only a small amount of the information is of interest.  Using this change,
you can simultaneously run a post-processing algorithm that boils output
down to just the part you are interested in and then deletes the original large
files without interfering with incomplete files that have the .part extension.
This avoids going over disk usage quotas, for example.