(AbstractAction.java) Added functions getAttribute and setAttribute for the IParticle in order to slowly get rid of the *ParticleLayer.java classes.
Also set public the Trackers getValue functions so that it can be called outside the io package, in order to slowly get rid of the Action specific *Tracker.java class 

Applied these changes to DebGrowth action. They must be validated though with a numerical experiment. It is not clear yet how to set the name of an attribute when we want it to be used throughout several *Action.java classes. For instance particle length.