Added new calendars : StandardCalendar.java is the default calendar with Julian dates before 1582 October 15th and Gregorian dates after. ProlepticGregorianCalendar.java (Gregorian dates before 1582 October 15th) JulianCalendar.java NoLeapCalendar.java, one year = 365 days always AllLeapCalendar.java, one year = 366 days always
Doing so I realised that InterannualCalendar.java behaves just like a ProlepticGregorianCalendar.java. This calendar and NoLeapCalendar and AllLeapCalendar that extend it, have a flaw though : they do not work for dates prior to the epoch date (== the user defined origin of time), whereas it should return negative time values. Will have to decide whether it is worth fix it up.